Update linker control files
There are two fixes done. First the data_init_table needs to be in
bank 00 when using the small code model. As this table typically is
small anyway, just put it in bank 00 for now. Second, the Scheme
interpreter used cannot load another file if the files are not in the
current directory. Copying the common definitions (back into) the
board specific files is a work around for now.