Rework interrupt.h
Having bit-fields turns out to not be the best idea. Accepting a
pending interrupt is done by writing a one which is the same value
as having a pending interrupt. This works nice for the bit(s) you are
working on, but trying to write this to the pending register as a
bit-field results in that all other pending bits are acknowledged.

It is much safer to use full register operations so that the developer
knows what is going on and this file is now completely reworked due to
this.