Move _InitialStack variable to a separate module
This is done to allow the standard C startup code to be used from the
C library. When downloading a hex file over the debug port it is
better to link with that C startup and having the variable defined in
the C256 C startup file caused linker issues as it got two C startups
linked in due to this.